+++ lib/parser.js
@@ -7,6 +7,7 @@ var functionHandlers = require("./handlers/functionHandlers");
 var invocationHandlers = require("./handlers/invocationHandlers");
 var moduleHandlers = require("./handlers/moduleHandlers");
 var commentHandlers = require("./handlers/commentHandlers");
+var propertyAccessorHandlers = require("./handlers/propertyAccessors");
 
 var parserState = require("./parserState");
 const StateEnum = parserState.StateEnum;
@@ -479,33 +480,6 @@ function handleVery(parseContext) {
 }
 
 /**
- * Handles property accession with an index:
- *  <object> levl <key>
- *
- * Produces:
- *  object[key];
- */
-function handleLevl(parseContext) {
-  tokenUtils.expectToken("levl", parseContext);
-
-  var tokens = parseContext.tokens;
-
-  // consume levl
-  tokens.shift();
-
-  // there should be a next token
-  if (tokens.length < 1) {
-    throw new SyntaxError(
-      `Expected argument but got nothing. Allowed construct: obj levl [arg]. ${parseInfo(
-        parseContext
-      )}`
-    );
-  }
-
-  return `[${tokens.shift()}]`;
-}
-
-/**
  * Handles explicit return:
  *  amaze <expression>
  *
@@ -717,13 +691,11 @@ function parseStatement(parseContext) {
   }
 
   if (tokens[0] === "giv") {
-    // consume giv
-    tokens.shift();
-    return ".";
+    return propertyAccessorHandlers.handleGiv(parseContext);
   }
 
   if (tokens[0] === "levl") {
-    return handleLevl(parseContext);
+    return propertyAccessorHandlers.handleLevl(parseContext);
   }
 
   if (tokens[0] === "amaze") {
